replace 'input' with 'data'
I am thinking that this may be a good thing to have testing set up for. I carefully used find and replace, so I'm confident that this will be fine, but I changed basically all of the files a little bit. I could submit them all on rosalind again, but this seems to me to be why testing exists.

fixes #25